Upgrade react version, reimplement example
