+++ SwipeListExample/package.json
@@ -3,11 +3,22 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start"
+    "start": "node node_modules/react-native/local-cli/cli.js start",
+    "test": "jest"
   },
   "dependencies": {
-    "react": "^15.2.0",
-    "react-native": "^0.28.0",
+    "react": "15.3.2",
+    "react-native": "0.36.1",
     "react-native-swipe-list-view": "file:../"
+  },
+  "jest": {
+    "preset": "jest-react-native"
+  },
+  "devDependencies": {
+    "babel-jest": "16.0.0",
+    "babel-preset-react-native": "1.9.0",
+    "jest": "16.0.2",
+    "jest-react-native": "16.1.0",
+    "react-test-renderer": "15.3.2"
   }
 }
